+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Requires PHP `8.5`
 - Requires `innmind/black-box:~7.0`
 
 ## 5.1.0 - 2026-03-21